chore: Migrate release pipeline to release-please
Migrated the release pipeline from a custom workflow to Google's release-please action.

This change updates the workflow to trigger on pushes to the `main` branch, and uses the release-please action to automatically create releases, generate release notes, and upload artifacts. Additionally, updated JDK version to 21 for building the Connector Jar.